garden-app_t/requirements.txt
streamlit
pillow
requests
openai
pandas
supabase-py==0.2.7
